+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## Unreleased
+
+## v0.4.1 - 2025-12-14
 - Dependencies: Added compatibility with docutils 0.19 - 0.22
 - Dependencies: Added compatibility with sphinx 7 - 9